requirements.txt
# H&M Data Assistant - Required Dependencies

# Core Framework
flask==2.3.3
flask-cors==4.0.0

# Azure Services  
azure-storage-blob==12.19.0
azure-identity==1.14.0

# Google Cloud / BigQuery
google-cloud-bigquery==3.13.0
google-auth==2.23.3

# OpenAI / Azure OpenAI
openai==1.3.7
azure-openai==0.1.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Environment & Configuration
python-dotenv==1.0.0

# Utility Libraries
requests==2.31.0
pydantic==2.4.2
typing-extensions==4.8.0

# Development Dependencies (optional)
pytest==7.4.3
black==23.9.1
flake8==6.1.0

# LangGraph for workflow management
langgraph>=0.0.55

# System monitoring (for health checks)
psutil==5.9.6

# Security libraries (for future auth implementation)
pyjwt==2.8.0
cryptography==41.0.7